Implement "natural order sort" in for coil names in CoilGroup (#657)
* Added test of sort-order of coils

The coils are being sorted by type and by name, with name being in
strictly alphabetical order. This means that single-digit
numbers (e.g. 7) can be sorted to come after two-digit numbers (e.g.,
11). See #655.

* Fixed sorting of coils by name

As described in #655, when the CoilSet class sorts coils it does so
using a naive alphabetical sort on the name. However, this fails to
handle numbers within the string, for example regarding 8 > 10. This
commits implements "natural sort order" to rectify that.