Move arguments to init (#650)
* Move arguments to init

This moves various kwargs usage, which is currently prevelent in many
function signatures, to the init function. In this way, each builder can
be set up and reinitialised as required, but is always called the same
way. The usage of explicit arguments on the implementations of the init
functions for concrete classes should also make it easier to follow what
arguments need to be passed in.

* Ensure equilibrium is handled correctly

* Call builders correctly in tests

* Handle builds when separatrix isn't passed in

Note that needs additional validation as we need a separatrix in
runmode=run.

* Assigned boundary when analysing equilibrium