Refactor build stage (#624)
* Refactor TFCoil so optimisation done by dependency injection

* Introduce callbacks to perform optimisation within reactor build

This allows the optimisation to be turned off, even if it is heavily
mixed into the build procedure.

* Fixed failing unit tests

Some tests failed because a previous commit had introduced a typo into
one module. Others failed because I renamed the `optimise` method for
TF coils. In the latter case I updated the name of the method in the
tests and made sure to pass in the callback.

* Add unit test for refactor of TF coil optimisation

This is a simple test which checks whether running without the
optimisation callback gives different results from running with it. It
also checks that running the build method a second time, that time
without the optimisation call, does not change the state of the TF
coil object.

* Refactor optimisation of first wall

The optimisation step of building the first wall system has been moved
into an optional callback. However, I'm not certain if this really
counts as optimisation. (I.e., is it actually changing any
parameters?) I will need to confirm that. I've also added a unit test
to make sure you get different results calling the build method with
and without the callback.

* its just CustomProfile tht was the problem

* remove to_dict scaling

* Add option to perform no PROCESS run.

* Import callbacks for BLUEPRINT reactor class

* Removed unused ATEC optimisation callback

* Added test of sort-order of coils

The coils are being sorted by type and by name, with name being in
strictly alphabetical order. This means that single-digit
numbers (e.g. 7) can be sorted to come after two-digit numbers (e.g.,
11). See #655.

* Fixed sorting of coils by name

As described in #655, when the CoilSet class sorts coils it does so
using a naive alphabetical sort on the name. However, this fails to
handle numbers within the string, for example regarding 8 > 10. This
commits implements "natural sort order" to rectify that.

* Made compare_dicts more flexible

I made it possible to specify the tolerance when doing an
"almost_equal" comparison. I also made it treat lists as though they
are arrays. This is useful when comparing a dict which has just been
read in from JSON.

* Fixed flake8 style issues and added to docstring info

* Minor changes requested by @je-cook in #624

* Addressed code review on PR #624

Co-authored-by: Matti <matti.coleman@ukaea.uk>