Perform wall convex hull in first wall builder
The MinimiseLength optimiser is quite general, so I don't think it
should be performing a convex hull over a list of keep out zones.
We may, for example, wish to pass it a list of independent keep out
zones at some point. The builder now has the responsibility to generate
the keep out zone and pass that to the optimiser.