Remove run/mock methods from DivertorBuilder
As the builder is not currently solving a design problem, it does not
require these methods.

I don't particularly like this method of popping the runmode from the
config in the parent builder before passing it down to the builders
for its components. We should look for a neater, less manual, way of
making this work.