Write initial first wall builder (#680)
* Add basic builder for EUDEMO first wall

The wall is currently a simple PolySpline in the xz plane, with no
optimisations implemented yet.

* Add test for FirstWallBuilder height calculation

* Add tests that components added to EUDEMO reactor

* Replace magic strings in EUDEMOReactor

* Begin FullFirstWallBuilder class

This class will build the first wall shape (using existing builder),
cut that shape below the x-point, build the divertor (builder to be
written), then join the wall and divertor.

In this commit, the cutting of the wall shape has been implemented.

* Split first wall builders into separate files

* Add test that first wall is closed shape

* Add basic target components in DivertorBuilder

The targets are drawn vertically rather than using a grazing angle to
determine the orientation of the targets.

* Add a few tests for DivertorBuilder

* Add dome to DivertorBuilder

* Rename Leg enum in divertor.py

* Fix TooManyArguments error in EUDEMO builders

After a rebase, parameters are no longer passed into the __call__ dunder
in Builder objects. Fix by simply removing the parameters from the call

* Add start and end_point methods to BluemiraWire

* Implement logic to join divertor to first wall

* Make FirstWallBuilder inherit from Builder

It does not need to inherit from ParameterisedBuilder (at least for now)

* Rename ClosedFirstWallBuilder -> WallBuilder

* Fix typo base config

* Fix mislabelled baffle in EUDEMO divertor builder

* Remove magic strings from first wall builders

* Add top-level component in FirstWallBuilder

* Add x_lims argument to DivertorBuilder

We previously had a start and end point, but the z-value was ignored
anyway as the x-point's z coord was used.

* Add optimised design problem to WallBuilder

* Define run and mock for DivertorBuilder

* Remove debugging variable and unused imports

* Update first wall component labels

* Clear up first wall component tree dimensionality

Previously, the tree looked something like this:
"first_wall" -> xz -> divertor -> xz -> ...

Having the nested "dimensionality" components does not make much sense,
and will be incorrect if components for different dimensions are added
to, e.g., the divertor

* Further tidy first wall component tree structure

* Add 'First Wall' component to EUDEMOReactor

* Resolve some flake8 warnings

* Add flake8 camel_case func name exception in tests

* Add test for first_wall component tree

* Use LCFS for first_wall exclusion zone

* Fix formatting in error message

* Add PrincetonD first wall shape to EUDEMO builder

* Refactor make_keep_out_zone in FirstWallBuilder

* Use PrinctonD for EUDEMO wall shape by default

This shape is much simpler than the PolySpline used previously. The
optimzer is finding it difficult to get a sensible result using
PolySpline, so use PrincetonD until we can work out the magic
combination of parameters/better shape to use.

* Remove hard-coded WallBuilder parameters

* Raise error if x-point outside of first wall

* Tweak WallPolySpline parameters

* Add missing space in parameterisation error

* Tidy DivertorBuilder class

* Add test for WallBuilder with PrinctonD

* Set # points in discretisation in MinimiseLength

* Perform wall convex hull in first wall builder

The MinimiseLength optimiser is quite general, so I don't think it
should be performing a convex hull over a list of keep out zones.
We may, for example, wish to pass it a list of independent keep out
zones at some point. The builder now has the responsibility to generate
the keep out zone and pass that to the optimiser.

* Move MinimiseLength into optimiser module

* Fix flake8: missing docstring and unused import

* Fix first wall config in EUDEMOReactor

* Remove unused return value in geometry tool

* Exit early on error in msh2xdmf install script

* Resolve TODOs in EUDEMO wall builders

* Set gradient in MinimiseLength objective function

* Fix return type in 'discretize()' type annotation

* Move first wall component building out of init

* Refactor keep-out zone func in first wall builder

* Remove unused import

* Replace relative imports with absolute

* Remove run/mock methods from DivertorBuilder

As the builder is not currently solving a design problem, it does not
require these methods.

I don't particularly like this method of popping the runmode from the
config in the parent builder before passing it down to the builders
for its components. We should look for a neater, less manual, way of
making this work.

* Add more detail to make_dome docstring

* Update make_baffle docstring

* Init component builders in FirstWallBuilder init

* Fix missing return after rebase

* Add component trees to first wall builders

* Remove unused import in first_wall

* Put component trees in docstring code block

* Remove kwargs from first wall builders

* Add convex_hull_wires_2d to geometry tools

* Remove duplicated convex hull function in builder

* Remove out-of-sync EUDEMOReactor docstring

The docstring listed the components built, this was already out-of-sync
with the actual components being built, and the code is explicit about
what is being built.

* Return Coordinates from BluemiraWire.start_point

And end_point. This is more consistent with methods like discretize,
which return Coordinates, not raw numpy arrays

* Move Wire start/end_point impls to _freecadapi

* Add missing asserts to test_freecadapi

* Fix flake8 errors

* Unused imports
* Length of section underline in docstring

* Fix numpy array orientation in test_tools test