Implement variable thickness offset function (#856)
* Add find_clockwise_angle_2d function to geom tools

This will be useful for implementing the variable offset/thickness
curve function required to build reactor blankets.

* Use degrees not radians in find_clockwise_angle_2d

Public functions should use degrees not radians.

* Implement varied_offset_function for BluemiraWire

* Implement varied_offset_function using normal vecs

* Fix incorrect outboard offset in varied_offset

* Refactor varied_offset_function into smaller funcs

* Refactor repeated code in test_varied_offset

* Rename varied_offset_function -> varied_offset

We know it's a function...

* Add variable offset end angle in varied_offset

* Update varied_offset tests

* Use make_bspline in varied_offset function

* Add docstring to varied_offset function

* Add docstrings to varied_offset private funcs

* Do not sort angles in varied_offset calculation

The sorting was mostly for my sanity when debugging as well as
making interpolating easier. We're no longer interpolating or debugging
so this sorting is not necessary

* Rename varied_offset arguments

* Tighten tolerance in varied_offset test

* Sort angles in varied_offset function

This ensures that coordinates are ordered in the clockwise direction
which makes it easier to find the outward facing normals to the tangents

* Add another shape to varied_offset test cases

* Make comment and docstring clarifications

* Remove unused function in varied_offset

* Add module docstring to varied_offset

* Add note about plane in varied_offset docstring

* Raise in varied_offset if wire is not closed

* Raise in varied_offset if wire is not planar in xz

* Rename arguments in varied_offset

The function is going to be used specifically for working out breeding
blanket thicknesses, so is going to be used in the context of a
tokamak's toroidal cross-section. Hence, inboard/outboard are more
descriptive names for the input variables

* Rename varied_offset params _angle -> _degrees

This is more consistent with conventions in the rest of the code base

* Move varied_offset into builders/EUDEMO

The function is quite specific to the EUDEMO design, and the builders
will be the only things using the function.

* Add input validation to varied_offset function

* Move varied_offset to builders/EUDEMO/tools

* Move varied_offset to builders/EUDEMO/tools

This was the suggested place for the function to live. My mistake
creating a new file in tools/first_wall