Fix contour generation bug caused by matplotlib
We use a class in an undocumented, private module in matplotlib.
The `matplotlib._contour` had a recent change which changed the API of
the QuadContourGenerator.create_contour function, so that it returns two
arrays, instead of one. The fix is to only return the first array from
our API; this is the array that was previously returned by matplotlib.

Link to relevant matplotlib change:
https://github.com/matplotlib/matplotlib/commit/178012d136b2a34e22489658108d4de685150a36