Fix recursion error in BluemiraPathPatch3D
Using an attribute that is undefined in __getattr__ will cause a
recursion error.

My guess is that we were getting away with this before as __getattr__
wasn't being called by Matplotlib's Patch constructor. However, after
updating Matplotlib from 3.3.4 to 3.5.1, we're no longer getting lucky.

Fix is to define the PathPatch before we call the Patch constructor.