Update matplotlib to 3.5.1 (#998)
* Bumpy matplotlib to v3.5.1

* Fix contour generation bug caused by matplotlib

We use a class in an undocumented, private module in matplotlib.
The `matplotlib._contour` had a recent change which changed the API of
the QuadContourGenerator.create_contour function, so that it returns two
arrays, instead of one. The fix is to only return the first array from
our API; this is the array that was previously returned by matplotlib.

Link to relevant matplotlib change:
https://github.com/matplotlib/matplotlib/commit/178012d136b2a34e22489658108d4de685150a36

* Fix BoP plotting test failure

The test failure was caused by an update to Matplotlib. The test
compares two figures, the figure generated by the newer matplotlib
version must have been slightly different (although I could not spot any
difference).

This was fixed by using Matplotlib's image comparison function, rather
than comparing the output file's data.

I've also used the tempfile module to perform clear-up of the generated
plots.

* Unpin Matplotlib version

* Add missing assert to BoP plotting test

* Fix recursion error in BluemiraPathPatch3D

Using an attribute that is undefined in __getattr__ will cause a
recursion error.

My guess is that we were getting away with this before as __getattr__
wasn't being called by Matplotlib's Patch constructor. However, after
updating Matplotlib from 3.3.4 to 3.5.1, we're no longer getting lucky.

Fix is to define the PathPatch before we call the Patch constructor.

* Fix Matplotlib deprecation warning

The deprecation warning was:

MatplotlibDeprecationWarning: Axes3D(fig) adding itself to the figure is
deprecated since 3.4. Pass the keyword argument auto_add_to_figure=False
and use fig.add_axes(ax) to suppress this warning. The default value of
auto_add_to_figure will change to False in mpl3.5 and True values will
no longer work in 3.6.  This is consistent with other Axes classes.

* Fix another Matplotlib deprecation warning

Deprecation warning was:

MatplotlibDeprecationWarning:
The modification of the Axes.patches property was deprecated in
Matplotlib 3.5 and will be removed two minor releases later. Use
Artist.remove() instead.

Use .remove() instead of allowing list.pop to call `del` on the patch