Refactor EUDEMO thermal shield design stages
Small change that preserves the normal signature of a "design_stage",
i.e., that it returns the component it designs/builds. The change also
means that components are only ever added to the component treee in the
'run' method - which makes it easier to keep track of what's been added.