Add GitHub workflow for opening PR develop->main
We want this to run every 6 weeks, which I do not think is possible
using schedules' 'cron' option. The solution here is a bit hacky:
every Monday morning cron will start the job, then, in bash, we check
whether today's date is a multiple of 42 days (6 weeks) after our first
release date.. If it is not, we skip the step that opens the pull
request.