Update release GitHub actions workflow
It now checks out main and performs a fast-forward-only merge with
develop. Then pushes that to main.

This strategy was chosen as GitHub does not allow a fast-forward merge
through pull requests (why though?!). We want the history of develop and
main to line up perfectly. This way we don't run into conflicts, and
it's possible to locate the tags in the history of both branches.