Migrate PROCESS solver to new interface (#1146)
* Process api additions (#1055)

* ✨ process species additions

* 🎨 improvements spurred by comments

* ✅ add a test

* 🐛 pass if process not installed

* 🐛 pass if process not installed

* Update bluemira/codes/process/api.py

Co-authored-by: Harry Saunders <33317174+hsaunders1904@users.noreply.github.com>

Co-authored-by: Harry Saunders <33317174+hsaunders1904@users.noreply.github.com>

* Fix PROCESS enum value typo CYRO -> CRYO

* Add refactored Setup class for PROCESS solver

* Add refactored Run class for PROCESS solver

* Remove unused attribute in test class

* Clean up logging in plasmod Run class

This logging is more consistent with what's in PROCESS, solver and more
specific about what's happening.

* Write tests ahead of PROCESS Teardown refactor

* Make PROCESS solver's Teardown tests pass

* Add refactored PROCESS solver class

* Remove old PROCESS solver implementation

* Fix typo in test that was causing it to pass

* Correct docstring in PROCESS solver

* Remove trailing commas from PROCESS test IN.DAT

Parameters whose values contained a trailing comma could not be read by
PROCESS. The file format is no longer supported by PROCESS, so we no
longer support it either. Remove the trailing commas so the IN.DAT file
is usable in tests.

* Fix PROCESS test not stopping mock.patch

This was causing the patched function being called in subsequent tests!
Whoops

* Add tests for PROCESS run modes

* Fix some find-replace errors

* Make solver's Teardown testable without PROCESS

Mock out the PROCESS file reader so that the Teardown class's logic can
be testing without PROCESS being installed.

* Catch case where PROCESS result missing param

PROCESS does not necessarily output all the values we need for a radial
build. There are certainly some MFiles which are missing some
parameters.

Catch these cases and raise a better error.

* Resolve PROCESS solver TODO

The below was not needed. The files are not required to run the solver.

* Fix unused variable in PROCESS solver

* Remove check for empty file

This will cause an error later on anyway, so the check is not needed.

* Fix misbehaving PROCESS tests

One of the tests was passing, but because an error of a similar type was
caught.

Check the error message to make sure the right error is thrown.

* Add missing docstrings to PROCESS Teardown

* Remove unused constant in test

* Replace old codes.interface module

* Restore 'plot_radial_build' to PROCESS API

* Add run_mode_cls property to Solver

This allows us to access the available run modes for a solver, without
prior knowledge of which Solver class it is.

* Update codes.wrapper to use new Solver interface

* Add __all__ imports to codes.plasmod

* Use run and read directories in PROCESS Teardown

* Fix file paths in PROCESS system tests

* Update codes.interface tests

* Add some missing docstrings in PROCESS solver

* Fix flake8 warning in base.design

* Fix test that fails for long git branch names

The look_and_feel.print_banner test fails if the git branch is too long.
The long name caused a new line in the banner, which made the asserted
banner length incorrect.

* Restore species methods to PROCESS solver

These had been removed as part of the refactor of the PROCESS API.

* Fix mocking of PROCESS module

The mock was overwriting the PROCESS module, instead of patching it.
The effect of this was causing a subsequent test to fail.

* Remove superfluous comment

* Use walrus operator to summarise if statement

* Remove unused imports

Co-authored-by: je-cook <81617086+je-cook@users.noreply.github.com>