Fix failing plotting tests (#1170)
* Fix AttributeErrror in TestIbeam

A plotting test was using an old API and calling '.plot' to plot a face.

Use 'plot_2d' instead.

* Fix compound deflection test plotting error

There is no 'max_deflection' attribute on Element. I believe the
intention was 'max_displacement'.

Also, the color_normer is not used if there is no stress i.e., in the
DeformedGeometryPlotter. So check that the normer exists before using
it.