Clean up untracked files after tests are run (#1176)
* Write mesh data to ignored dir in equilibria test

The data was previously being written to the home directory, which was
cluttering things a bit. One of the output files (Mesh_linkfile.json)
was not ignored, so was cluttering Git's untracked files list.

* Fix typo inizialized -> initialized

* Add Mesh_linkfile.json to git ignore

This is an output file of some of our meshing routines.

* Remove test generated JSON from repo

* Do not write file in numpy JSON encoder test

Use dumps and loads, to write to strings instead. This is quicker and
prevents another test generated file from polluting the repo.

* Avoid test mesh files being written to cwd

* Remove unused import