Add warnings to CI and run more tests for the develop dependencies PRs (#1180)
* Run longrun tests in CI if merging dependencies

We can be more thorough here as merging of dependencies is done less
often.

* Add script to print warning reports

This will be used to display warnings on the CI.

* Add ability to compare warning reports

* Report new Python warnings in CI builds

Makes changes such that, after the CI runs on a PR, a comment is made on
the PR that lists any new warnings. New warnings do not fail the build,
these are just notifications.

* Fix PR branch reference in ci.yml

* Remove redundant 'if' in warning report script

* Improve report script's docs

* Use fppf-bot token for API auth

We were hitting rate limit errors using the repo's auth token, so let's
try using the bot account.

* Correct syntax for accessing GH actions secrets

* Use GitHub's token, not fppf-bot's, in CI

* Use GitHub API to download artifact

Using the GitHub Action workflow was maxing out our API request limit
because (I assume) it was searching through all previous builds to look
for our artifact.

We can use the GH API to do this search using a single request.