Disable Matplotlib Agg warning in tests
Matplotlib raises a warning if it's using a non-interactive backend and
plt.show() is called.

Disable this in pytest, as we are expecting not to see the plots if
we're not running tests with '--plotting-on'.

I tried disabling the warning within conftest.py only if plotting was
disabled, but pytest still picks up the warnings for some reason. Adding
it to the pytest config was the only way I could get the warnings
ignoredd.