Mock FreeCAD API viewer in tests, not Qt exec
Mocking QAppilication.exec_ was causing an error in the GitHub Actions
CI:

qt.qpa.xcb: could not connect to display
qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

I'm still not clear what caused this. Let's try to get around the issue
for now...