Run plotting tests with non-interactive backend (#1294)
* Remove 'if test.PLOTTING' conditions from tests

We ran into problems of plotting tests starting to fail as they were not
run on the CI. The plan is to use a non-interactive backend, so we can
verify that the plotting works as expected, without interactive windows
blocking test execution.

* Disable CAD viewer if not 'plotting-on' in tests

* Disable Matplotlib Agg warning in tests

Matplotlib raises a warning if it's using a non-interactive backend and
plt.show() is called.

Disable this in pytest, as we are expecting not to see the plots if
we're not running tests with '--plotting-on'.

I tried disabling the warning within conftest.py only if plotting was
disabled, but pytest still picks up the warnings for some reason. Adding
it to the pytest config was the only way I could get the warnings
ignoredd.

* Fix Matplotlib '> 20 figs open warning' in tests

Matplotlib was displaying a warning when the whole test suite was run,
saying that there are more than 20 figures open.

Fix this warning by explicitly closing figures in the tests.

* Remove unused import in test_reactor

* Remove TODO that's not required

The test intentionally mocks out the plotting, so we would not expect
the figures to be shown

* Mock FreeCAD API viewer in tests, not Qt exec

Mocking QAppilication.exec_ was causing an error in the GitHub Actions
CI:

qt.qpa.xcb: could not connect to display
qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

I'm still not clear what caused this. Let's try to get around the issue
for now...

* Remove longrun flag on fast test

* Close some plots left open in tests

Too many of these plots being open (> 20) triggers a Matplotlib warning.

* Remove extra plt.show call in test_transformation

* Add plt.show call to TestSankey

This means the figure will be shown when tests are run with
--plotting-on