Re-define Designer interface
This aims to solve the problem of having read/run/mock modes in the
designer, and determining which mode to run depending on the
build_config provided.

The Designer is now intended to run independently of a Builder, and acts
as a general Bluemira "task". There are some parallels with the `Task`
class within the `solver` module - these are to be resolved later.