Update ParameterFrame usage in codes module (#1489)
* Add ParameterFrame class for PlasmodSolver

* Remove old ParameterFrame from Plasmod solver

* Replace ParameterFrame usage in process module

This is pretty ugly at this point; needs tidying up quite a bit, but the
tests pass!

* Update radial build parameters in process solver

* Fix PROCESS output params not being updated

* Remove unused codes module functions

* Fix get_raw_outputs in process solver

* Add missing process parameter docstrings

* Fix plasmod solver

* Add MappedParameterFrame to codes module

* Use MappedParamterFrame in codes solvers

* Uncomment temporarily removed code

I think we should probably remove it permanently at some point. We
should try to avoid commonly used/public functions that can return more
than one type.

* Remove redundant comments

* Add test for obsolete input var in process solver

* Add test for parameter frame iterator

* Remove redundant __eq__ method from ParmeterFrame

This method is automatically generated by 'dataclass'.

* Add some missing ParameterFrame tests

* Remove unused Parameter.to_dict argument

* Remove unused import

* Remove some more unused imports

* Fix missing backtick in MappedParameterFrame doc

* Do not catch AttributeError in modify_mappings

* Use dict comprehension instead of loop

* Add tests for codes parameter mappings

We should check that all parameters have mappings, or we're going to run
into issues. This should be useful as a reminder to add a mapping when a
parameter is added.

* Inherit MappedParameterFrame in Plasmod parameters

Clean up some docstrings as well

* Remove unused function in PROCESS solver

This functionality was moved to the ProcessSolverParams class

* Remove docstring for deleted parameter

* Handle obsolete vars better in PROCESS solver

Make sure we can handle when the old variable is replaced by a list of
variables.

* Fix bug in process solver's teardown tests

Changes to the FakeMFile's data were being persisted across all the
tests. This was obviously causing issues. I'm not 100% happy with the
way I've fixed this, but it works as it needs to.