Replace uses of old Parameter and ParameterFrame with new versions (#1505)
* Remove old parameter module

* Use new ParameterFrame class in BoP module

* Remove parameter, and wildcard, imports from base

* Delete Configuration class

With the new ParameterFrame class, this is no longer needed.

* Replace usage of old-style params in TF builder

* Update docstring in OptVariables

* Re-define removed ParmeterMapping class

This was prematurely removed with the base.parameter module. Re-define
it where it is used, within the codes module.

* Remove commented test code

This was just a copy of a defined test that accidentally got left lying
around.

* Fix ParamteterFrame/Config import errors

A rebase meant some imports that worked on the base branch were no
longer working on this branch, as the modules being imported were
moved/removed.

* Remove ParameterFrame from balance of plant model

* Fix apostrophe in docstring

* Rename NewParameter[Frame] -> Parameter[Frame]

Now the old versions of the classes have been removed, we can rename the
new ones.

* Fix 'import x as x' occurrences

* Fix type annotation in _frame

* Fix ParameterFrame usage in TF builder/designer

* Remove 'wrapt' dependency

This was being used by the old Parameter class, but is no longer used in
the new version of the class.

* Remove function from Numpy JSON encoder

The function's purpose was to deal with the old Parameter objects, which
used 'wrapt' so they looked like floats, but actually weren't. This
confused the JSON encoder, hence this workaround. We no longer use
'wrapt' in the new version of the class, so this can be removed.

* Fix maintenance test in EU-DEMO design example

* Remove unused JSON encoder functions

Related to 53554b22