Use in-place bluemira install in examples CI
My theory is that doing a normal pip-install installs bluemira into
site-packages, but does not copy the examples with it. Hence any
examples that use `get_bluemira_path("", "examples")` are failing
because it's looking in the wrong place for the examples.

Using an in-place install means that the bluemira module is not moved
relative to the examples, so (hopefully) the 'get_.._path' call will
work.