Run examples in CI (#1748)
* Add script to run examples

* Add GitHub action to run examples

* Add 'push' to examples trigger

This is just to get the workflow triggered once, so we can trigger it
manually from then on

* Revert "Add 'push' to examples trigger"

This reverts commit a00b826a50b6279db462a0eaa2e9e36f90359e71.

* Use bash login shell when running examples in CI

* Pip install bluemira in examples CI

* Close plots after each example in examples script

* Print summary of results in 'run_examples'

* Disable process/plasmod dependent examples in CI

* Flesh out examples CI script docstring

* Fix 'run_examples' script CLI argument help

* Use in-place bluemira install in examples CI

My theory is that doing a normal pip-install installs bluemira into
site-packages, but does not copy the examples with it. Hence any
examples that use `get_bluemira_path("", "examples")` are failing
because it's looking in the wrong place for the examples.

Using an in-place install means that the bluemira module is not moved
relative to the examples, so (hopefully) the 'get_.._path' call will
work.

* Flush stdout after each example in examples script

In the CI logs, some error messages were being printed under the wrong
example's banner. I'm hoping flushing after each example is run will fix
this.

* Add comment on disabled examples in examples CI

* Add 'py' extension to examples script

This keeps consistency with the other Python scripts

* Remove redundant 'except' case

* Add typing to examples script function arg

Co-authored-by: je-cook <81617086+je-cook@users.noreply.github.com>

Co-authored-by: je-cook <81617086+je-cook@users.noreply.github.com>