Install dependencies from source (#701)
* Some rough WIP scripts

* Updates

Set compile flags to fix include and linking issues.
Need to do something about LD_LIBRARY_PATH to include files in
site-packages... best I've been able to work out so far is to hack the
activate script.

* Add more into LD_LIBRARY_PATH

* Copy across all Part libraries

* Install and build for pyside2-tools

* Install specific modules

* Install as per website

* Update scripts

* pyside2-tools handled by PySide2 build

* Update PySide2 paths

* Install pivy from source

* Install OCC

Work in progress - need to find a commit that works!

* Remove unused scripts

* Rename OCC script and try new commit

* Need an OCC dependency

* Update install sequence

* Actually do the install

* Add parameters to control number of parallel jobs

Plus some adjacent fixes.

* Specific dependencies for coin and pivy

* Update README following first run through

* Update README to install dependencies

* Remove libcoin dependency

* Align formatting

* Add a force rebuild flag

* Add instructions to install dependencies

* Add pyside2 build step back in

* Don't use sudo in scripts

* First pass at dockerfile

* Make dev and env images

Ideally we'd only copy across the runtime dependencies as the current
approach results in a very large image. Potentially some work needed to
find what would need to be copied across in that case.

* Make sure errors in scripts are propagated up

* Update apt-get requirements

* Update requirements in dockerfile

* Add OCC build back in

* Change ownership of copied directory to user

* 🚧 some hacks for ubuntu 22.04

* 🚧 separate freecad pulling and building

* 🚧 separate pyside pulling and building

* 🚧 separate qt5 steps

* 🚧 add fenics

* 🚧 try and fix pivy

* 🚧 Update Dockerfile

* 🐛 patch pivy

* 🐛 patch files

* ✨ docker compose files

* 🐛 remove spaces

* 🎨 one compose file

* ⚡️ less hacks

* ⚡️ use ninja

* 🐛 docker compose fixes

* 🐛 docker compose fixes

* 🎉 No Ascii failure

* 🚧 update hacks

* ⬆️ freecad 0.20

* 🔥 remove numba-scipy source install

* 🎨 meshio additional deps and rearrangement

* 🎨 remove hacks

* 🐛 fenics build fix

* 🎨 updates and shrink container

* ⬆️ pivy and freecad update

* 🐛 fix rebase

* 🔥 removed locales

* 🚧 pyside skip modules stuff

* Remove SKIPMODS arg from pyside build script

The pyside build failed complaining of missing Qt libraries. We may wish
to revisit this to avoid building more than we need to, but this makes
the build work.

* Add fenics build scripts for docker

Due to time constraints, we have not yet migrated from fenics to
fenicsx, which means we need fenics in this version of the docker
container.

Still required are steps for meshio and hdf5.

* Split up ENV commands in Dockerfile

This makes it easier to comment on the 2 paths being set

* Split docker build into base, release and develop

* Install 'main' not tag in Dockerfile

* Do Dockerfile copies within base image

* Revert changes to pre-commit config

Diff files are no longer around, so these changes aren't needed.

* Remove OCC install scripts

These are no longer required in the docker container, as we install what
we need via 'apt' and 'pythonocc' is no longer a dependency

* Add docker install script for h5py

meshio depends on h5py, but but a normal pip install of h5py resulted in
a segfault when it was imported by meshio (see code comment).

Using the system hdf5 libraries to build h5py works.

* Install non-dev version of coolprop in docker

* Add user_base image layer to Dockerfile

* Remove generic type from ParamDictT

This is not allowed since Python3.9:
https://bugs.python.org/issue43923

* Fix h5py install command in script

* Actually fix h5py pip install command

* Attempt to make STP test FreeCAD version agnostic

* Re-order Dockerfile image layers

Have 'release' at the bottom, so it's the one that gets tagged

* Remove out-of-date README for scripts dir

* Do not pip-install eudemo or examples

When pip installing, we were also installing the 'examples' and 'eudemo'
"packages".

Avoid this, as these are extras/shouldn't be installed

Co-authored-by: James <james.cook1@ukaea.uk>
Co-authored-by: Harry Saunders <33317174+hsaunders1904@users.noreply.github.com>